# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for AI Product Fit Scoring Engine
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface SignupBody {
  email: string;
  password: string;
  name: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface AuthUser {
  id: string;
  email: string;
  name: string;
  onboardingCompleted: boolean;
}

export interface AuthResponse {
  user: AuthUser;
  token: string;
}

export interface ApparelPreferences {
  topSize: string;
  bottomSize: string;
}

export interface ArMeasurements {
  /** @nullable */
  chest: number | null;
  /** @nullable */
  waist: number | null;
  /** @nullable */
  hips: number | null;
  /** @nullable */
  inseam: number | null;
}

export interface EmailIntegration {
  connected: boolean;
  categories: string[];
  brands: string[];
}

export interface UserProfile {
  userId: string;
  name: string;
  gender: string;
  /** @nullable */
  height: number | null;
  /** @nullable */
  weight: number | null;
  apparel: ApparelPreferences;
  arMeasurements: ArMeasurements;
  interests: string[];
  emailIntegration: EmailIntegration;
}

export interface UpdateProfileBody {
  name: string;
  gender: string;
  /** @nullable */
  height: number | null;
  /** @nullable */
  weight: number | null;
  apparel: ApparelPreferences;
  arMeasurements: ArMeasurements;
  interests: string[];
  emailIntegration: EmailIntegration;
}

export interface AnalyzeProductBody {
  url: string;
  boostedFeatures?: string[];
}

export interface BoostScoreBody {
  productUrl: string;
  boostedFeature: string;
}

export interface FeatureScores {
  /** @nullable */
  camera: number | null;
  /** @nullable */
  battery: number | null;
  /** @nullable */
  performance: number | null;
  /** @nullable */
  value: number | null;
  /** @nullable */
  design: number | null;
  /** @nullable */
  durability: number | null;
}

export interface FitScoreResult {
  productName: string;
  productUrl: string;
  category: string;
  fitScore: number;
  riskLevel: string;
  strengths: string[];
  weaknesses: string[];
  riskFactors: string[];
  whyItFitsYou: string[];
  whyItMayNot: string[];
  featureScores: FeatureScores;
  boostableFeatures: string[];
  analyzedAt: string;
}

export interface HistoryItem {
  id: string;
  productName: string;
  productUrl: string;
  category: string;
  fitScore: number;
  riskLevel: string;
  analyzedAt: string;
}

export type HistoryStatsCategoryBreakdown = { [key: string]: number };

export interface HistoryStats {
  totalAnalyzed: number;
  averageFitScore: number;
  /** @nullable */
  topCategory: string | null;
  categoryBreakdown: HistoryStatsCategoryBreakdown;
  /** @nullable */
  recentHighScore: number | null;
}