# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for AI Product Fit Scoring Engine
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Create a new user account
 */
export const SignupBody = zod.object({
  email: zod.string(),
  password: zod.string(),
  name: zod.string(),
});

/**
 * @summary Login with email and password
 */
export const LoginBody = zod.object({
  email: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  user: zod.object({
    id: zod.string(),
    email: zod.string(),
    name: zod.string(),
    onboardingCompleted: zod.boolean(),
  }),
  token: zod.string(),
});

/**
 * @summary Logout the current user
 */
export const LogoutResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Get current authenticated user
 */
export const GetMeResponse = zod.object({
  user: zod.object({
    id: zod.string(),
    email: zod.string(),
    name: zod.string(),
    onboardingCompleted: zod.boolean(),
  }),
  token: zod.string(),
});

/**
 * @summary Get the current user's profile
 */
export const GetProfileResponse = zod.object({
  userId: zod.string(),
  name: zod.string(),
  gender: zod.string(),
  height: zod.number().nullable(),
  weight: zod.number().nullable(),
  apparel: zod.object({
    topSize: zod.string(),
    bottomSize: zod.string(),
  }),
  arMeasurements: zod.object({
    chest: zod.number().nullable(),
    waist: zod.number().nullable(),
    hips: zod.number().nullable(),
    inseam: zod.number().nullable(),
  }),
  interests: zod.array(zod.string()),
  emailIntegration: zod.object({
    connected: zod.boolean(),
    categories: zod.array(zod.string()),
    brands: zod.array(zod.string()),
  }),
});

/**
 * @summary Create or update the user's profile (called after onboarding)
 */
export const UpdateProfileBody = zod.object({
  name: zod.string(),
  gender: zod.string(),
  height: zod.number().nullable(),
  weight: zod.number().nullable(),
  apparel: zod.object({
    topSize: zod.string(),
    bottomSize: zod.string(),
  }),
  arMeasurements: zod.object({
    chest: zod.number().nullable(),
    waist: zod.number().nullable(),
    hips: zod.number().nullable(),
    inseam: zod.number().nullable(),
  }),
  interests: zod.array(zod.string()),
  emailIntegration: zod.object({
    connected: zod.boolean(),
    categories: zod.array(zod.string()),
    brands: zod.array(zod.string()),
  }),
});

export const UpdateProfileResponse = zod.object({
  userId: zod.string(),
  name: zod.string(),
  gender: zod.string(),
  height: zod.number().nullable(),
  weight: zod.number().nullable(),
  apparel: zod.object({
    topSize: zod.string(),
    bottomSize: zod.string(),
  }),
  arMeasurements: zod.object({
    chest: zod.number().nullable(),
    waist: zod.number().nullable(),
    hips: zod.number().nullable(),
    inseam: zod.number().nullable(),
  }),
  interests: zod.array(zod.string()),
  emailIntegration: zod.object({
    connected: zod.boolean(),
    categories: zod.array(zod.string()),
    brands: zod.array(zod.string()),
  }),
});

/**
 * @summary Analyze a product URL and generate a personalized Fit Score
 */
export const AnalyzeProductBody = zod.object({
  url: zod.string(),
  boostedFeatures: zod.array(zod.string()).optional(),
});

export const AnalyzeProductResponse = zod.object({
  productName: zod.string(),
  productUrl: zod.string(),
  category: zod.string(),
  fitScore: zod.number(),
  riskLevel: zod.string(),
  strengths: zod.array(zod.string()),
  weaknesses: zod.array(zod.string()),
  riskFactors: zod.array(zod.string()),
  whyItFitsYou: zod.array(zod.string()),
  whyItMayNot: zod.array(zod.string()),
  featureScores: zod.object({
    camera: zod.number().nullable(),
    battery: zod.number().nullable(),
    performance: zod.number().nullable(),
    value: zod.number().nullable(),
    design: zod.number().nullable(),
    durability: zod.number().nullable(),
  }),
  boostableFeatures: zod.array(zod.string()),
  analyzedAt: zod.string(),
});

/**
 * @summary Recalculate fit score with boosted feature weight
 */
export const BoostScoreBody = zod.object({
  productUrl: zod.string(),
  boostedFeature: zod.string(),
});

export const BoostScoreResponse = zod.object({
  productName: zod.string(),
  productUrl: zod.string(),
  category: zod.string(),
  fitScore: zod.number(),
  riskLevel: zod.string(),
  strengths: zod.array(zod.string()),
  weaknesses: zod.array(zod.string()),
  riskFactors: zod.array(zod.string()),
  whyItFitsYou: zod.array(zod.string()),
  whyItMayNot: zod.array(zod.string()),
  featureScores: zod.object({
    camera: zod.number().nullable(),
    battery: zod.number().nullable(),
    performance: zod.number().nullable(),
    value: zod.number().nullable(),
    design: zod.number().nullable(),
    durability: zod.number().nullable(),
  }),
  boostableFeatures: zod.array(zod.string()),
  analyzedAt: zod.string(),
});

/**
 * @summary Get recent product analysis history for current user
 */
export const GetHistoryResponseItem = zod.object({
  id: zod.string(),
  productName: zod.string(),
  productUrl: zod.string(),
  category: zod.string(),
  fitScore: zod.number(),
  riskLevel: zod.string(),
  analyzedAt: zod.string(),
});
export const GetHistoryResponse = zod.array(GetHistoryResponseItem);

/**
 * @summary Get aggregate stats from user's scoring history (avg score, categories, top brands)
 */
export const GetHistoryStatsResponse = zod.object({
  totalAnalyzed: zod.number(),
  averageFitScore: zod.number(),
  topCategory: zod.string().nullable(),
  categoryBreakdown: zod.record(zod.string(), zod.number()),
  recentHighScore: zod.number().nullable(),
});