# lib/api-zod/src/generated/types/apparelPreferences.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for AI Product Fit Scoring Engine
 * OpenAPI spec version: 0.1.0
 */

export interface ApparelPreferences {
  topSize: string;
  bottomSize: string;
}