# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for AI Product Fit Scoring Engine
 * OpenAPI spec version: 0.1.0
 */

export interface AuthUser {
  id: string;
  email: string;
  name: string;
  onboardingCompleted: boolean;
}