# lib/api-zod/src/generated/types/emailIntegration.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for AI Product Fit Scoring Engine
 * OpenAPI spec version: 0.1.0
 */

export interface EmailIntegration {
  connected: boolean;
  categories: string[];
  brands: string[];
}