# lib/api-zod/src/generated/types/featureScores.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for AI Product Fit Scoring Engine
 * OpenAPI spec version: 0.1.0
 */

export interface FeatureScores {
  /** @nullable */
  camera: number | null;
  /** @nullable */
  battery: number | null;
  /** @nullable */
  performance: number | null;
  /** @nullable */
  value: number | null;
  /** @nullable */
  design: number | null;
  /** @nullable */
  durability: number | null;
}