# lib/api-zod/src/generated/types/fitScoreResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for AI Product Fit Scoring Engine
 * OpenAPI spec version: 0.1.0
 */
import type { FeatureScores } from "./featureScores";

export interface FitScoreResult {
  productName: string;
  productUrl: string;
  category: string;
  fitScore: number;
  riskLevel: string;
  strengths: string[];
  weaknesses: string[];
  riskFactors: string[];
  whyItFitsYou: string[];
  whyItMayNot: string[];
  featureScores: FeatureScores;
  boostableFeatures: string[];
  analyzedAt: string;
}