# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for AI Product Fit Scoring Engine
 * OpenAPI spec version: 0.1.0
 */
import type { ApparelPreferences } from "./apparelPreferences";
import type { ArMeasurements } from "./arMeasurements";
import type { EmailIntegration } from "./emailIntegration";

export interface UserProfile {
  userId: string;
  name: string;
  gender: string;
  /** @nullable */
  height: number | null;
  /** @nullable */
  weight: number | null;
  apparel: ApparelPreferences;
  arMeasurements: ArMeasurements;
  interests: string[];
  emailIntegration: EmailIntegration;
}